:truck: renamed to HostTest and HostTestFactory
