+++ src/Codebelt.Extensions.Xunit.Hosting/Internal/HostTest.cs
@@ -2,30 +2,30 @@
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
-namespace Codebelt.Extensions.Xunit.Hosting
+namespace Codebelt.Extensions.Xunit.Hosting.Internal
 {
-    internal sealed class GenericHostTest : HostTest<IHostFixture>
+    internal sealed class HostTest : HostTest<IGenericHostFixture>
     {
         private readonly Action<IServiceCollection> _serviceConfigurator;
         private readonly Action<HostBuilderContext, IServiceCollection> _serviceConfiguratorWithContext;
         private readonly Action<IHostBuilder> _hostConfigurator;
         private HostBuilderContext _hostBuilderContext;
 
-        internal GenericHostTest(Action<IServiceCollection> serviceConfigurator, Action<IHostBuilder> hostConfigurator, IHostFixture hostFixture) : base(true, hostFixture, callerType: serviceConfigurator?.Target?.GetType() ?? hostConfigurator?.Target?.GetType())
+        internal HostTest(Action<IServiceCollection> serviceConfigurator, Action<IHostBuilder> hostConfigurator, IGenericHostFixture hostFixture) : base(true, hostFixture, callerType: serviceConfigurator?.Target?.GetType() ?? hostConfigurator?.Target?.GetType())
         {
             _serviceConfigurator = serviceConfigurator;
             _hostConfigurator = hostConfigurator;
             InitializeHostFixture(hostFixture);
         }
 
-        internal GenericHostTest(Action<HostBuilderContext, IServiceCollection> serviceConfigurator, Action<IHostBuilder> hostConfigurator, IHostFixture hostFixture) : base(true, hostFixture, callerType: serviceConfigurator?.Target?.GetType() ?? hostConfigurator?.Target?.GetType())
+        internal HostTest(Action<HostBuilderContext, IServiceCollection> serviceConfigurator, Action<IHostBuilder> hostConfigurator, IGenericHostFixture hostFixture) : base(true, hostFixture, callerType: serviceConfigurator?.Target?.GetType() ?? hostConfigurator?.Target?.GetType())
         {
             _serviceConfiguratorWithContext = serviceConfigurator;
             _hostConfigurator = hostConfigurator;
             InitializeHostFixture(hostFixture);
         }
 
-        private new void InitializeHostFixture(IHostFixture hostFixture)
+        private void InitializeHostFixture(IGenericHostFixture hostFixture)
         {
             if (!hostFixture.HasValidState())
             {
@@ -35,8 +35,7 @@ namespace Codebelt.Extensions.Xunit.Hosting
                 hostFixture.ConfigureHost(this);
             }
             Host = hostFixture.Host;
-            ServiceProvider = hostFixture.Host.Services;
-            Configure(hostFixture.Configuration, hostFixture.HostingEnvironment);
+            Configure(hostFixture.Configuration, hostFixture.Environment);
         }
 
         protected override void ConfigureHost(IHostBuilder hb)
@@ -51,7 +50,7 @@ namespace Codebelt.Extensions.Xunit.Hosting
             _serviceConfiguratorWithContext?.Invoke(Tweaker.Adjust(_hostBuilderContext, hbc =>
             {
                 hbc.Configuration = Configuration;
-                hbc.HostingEnvironment = HostingEnvironment;
+                hbc.HostingEnvironment = Environment;
                 return hbc;
             }), services);
         }
